Assignment1/requirements.txt
torch
torchvision
matplotlib
numpy<2
pillow
opencv-python
